Phpstan improvements (#1487)
* typehinting and import fixes

* lots more fixes

* lots more fixes

* lvl 4 fixes

* move stuff to baseline

* fixes

+++ src/Domain/Shared/Support/Throttling/SimpleThrottle.php
@@ -32,7 +32,7 @@ class SimpleThrottle
 
         $mailsPerSecond = config("mail.mailers.{$mailer}.mails_per_timespan", $this->allowedNumberInPeriod) / config("mail.mailers.{$mailer}.timespan_in_seconds", $this->periodLengthInSeconds);
 
-        $this->allowedNumberInPeriod = ceil($mailsPerSecond * 2);
+        $this->allowedNumberInPeriod = (int) ceil($mailsPerSecond * 2);
         $this->periodLengthInSeconds = 1;
 
         return $this;
